requirements.txt
pynmea2
pyserial
ssh2-python